services/user/requirements.txt
Authlib==1.3.2
babel==2.16.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.7
cryptography==43.0.3
esdk-obs-python==3.24.6.1
fakeredis==2.26.1
Flask==3.1.0
flask-babel==4.0.0
Flask-Mail==0.10.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pycryptodome==3.21.0
pytest==8.3.3
pytest-cov==6.0.0
python-dotenv==1.0.1
pytz==2024.2
pyuwsgi==2.0.28.post1
redis==5.2.1
requests==2.32.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
types-pytz==2024.2.0.20241003
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
